Use tls socket to retrieve distinguished name
Previously, it was possible to get DN using dn_callback passed to
credentials object. This callback is invoked on every handshake (if
'client_auth' is set to true) and accepts two parameters, subject and
issuer. But there is no way for the user of the tls library to connect
particular client connection with the DN string.

This commit adds a 'get_distinguished_name' method to connected socket
interface. This commit can be used to wait on a future that returns a
distinguished name or nullopt after a handshake. The handshake is
performed asynchrnously. Because of that the retrieval of the DN string
should also be asynchronous.

This commit also adds a unit test and scripts that generate a bunch of
certificates/keys for it.