Allow application to query the abort on OOM state
Add an is_abort_on_allocation_failure method which returns
the global state of this reactor option.

This is useful for applications that which to query this state in order
to expose it for debugging or administration purposes. Applications
could try to track this themselves, but this is unreliable since the
reactor itself can change this value based on the command line flags.