memory: Use separate small pool for tracking sampled allocations
To differentiate sampled from non-sampled allocations we use a separate
set of small pools.

Each pool gets tagged to indicate whether it's for sampled allocations
or not.

Hence, on free we can simply check whether an allocation belongs to a
sampled-pool or not and act accordingly.

For sampled pools we could reduce the amount of target free objects in a
pool but for any serious workload this probably doesn't matter.

Large allocations continue to use the `allocation_site` pointer on the
`page` structure. However, we now always need to check whether the
pointer is valid as it's null if not sampled.