core/cpu_profiler: avoid taking samples during exception unwinding
In libgcc there is a critical section where the stack is being modified so
execution can return to an exceptions landing pad. However, this
partially modified state isn’t valid or specified by the dwarf info in
the eh_frames. So a seg fault tends to occur when `backtrace()` tries to
unwind through this partially modified stack and follows an invalid
pointer in it.