net/tls: Adjust type for cert_info.serial
GnuTLS provides serial numbers not as a legible string of
hex digits but as a byte array corresponding to an integer
value up to 20 octets wide. The cert_info interface should
reflect this.

To that end, this commit updates the type of `cert_info::serial`
to an non-null-terminated basic_sstring of uint8_t.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>