aio_general_context: Allow more than max_poll() queued iocbs
The `aio_general_context` had the implicit assumption that in a single
tick we would never queue more than `--max-networking-io-control-blocks`
events/iocbs. This however ignores situations such as queuing multiple
iocbs per socket per tick, having left over iocbs in the queue from
previous ticks via the new EAGAIN handling or simply because a lot more
sockets are in use which isn't prevented anywhere else.

If this condition was hit (`last > end`) the reactor would just assert
out and crash.

To avoid this situation this patch introduces a backlog into which
elements are being enqueued when the original array is full and which
can grow unbounded. This mirrors how the `aio_storage_context` works
which uses the `io_sink` for the same purpose.

To avoid oversized allocations after startup the split into two separate
data structures is needed (instead of just regrowing the array). Further
the datastructure from which the iocbs are passed into `io_submit` needs to
be in contigiuous memory (and also provide an API to use it which most
containers don't).

`std::deque` is used in the backlog to avoid oversized allocations in
the backlog itself. The existing array solution for `iocbs` is kept to
fulfill the contigiuous memory requirement.

Further we slightly change how EAGAIN is handled. Instead of
backshifting the array we keep the array as is and just track the
`begin` of the array across `flush` calls. This is possible now as the
backlog handling is in place.

This introduces "batching" and prevents degenerate cases where only a
single element is being submitted which would then result in repeated
shifting of the whole array on each `flush` call. Given we use a chunked
data structure like `std::deque` erasing from the front of the backlog
is relatively cheap and does not require shifting all the elements in
the backlog. Hence, the per-iocb overhead is amortized constant.

Note that in general we try to submit as many iocbs per `io_submit`
call. Given the new behaviour of not backshifting the iocb array and
immediately backfilling from the backlog we might introduce `io_submit`
calls that don't try to submit the max amount of iocbs.

However we assume that if we ran into EAGAIN then either:

 - We are still behind the next time around: it's unlikely we would
   succeed in submitting all the iocbs anyway
 - We have now caugt up: we have introduced a single additional
   `io_submit` call which only submits `max_poll()/2` iocbs on average.
   The backlog will be drained at full `max_poll` per `io_submit`.

(cherry picked from commit d9175fcf84d31f21c0b72d0bea405f73ac1009b4)