core: make custom deleter constructor private
In the Redpanda fork, constructor a temporary buffer from a passed-in
delete is potentially unsafe as deleters
may be invoked on _any_ thread, so a deleter must be safe to call in
this way. In seastar upstream, temporary_buffer is not safe to share
across threads, so deleters are called in a more restricted way.

Make the constructor private and add a static factory method which
uses this constructor, so that we can audit all uses of this
constructor to ensure they are safe or that we do not call them.