Add scoped to fallback to system alloc
Add a scoped_system_alloc_fallback object, which changes the behavior
when at least one of these objects is in scope: when a large allocation
would otherwise fail, we instead fall back to the system allocator.

This would allow limited use of the system allocator in cases where it
is useful to do large allocations, which might otherwise fail due to
fragmentation.