tls: stub out force_rehandshake
Upstream implemented force_rehandshake on the GnuTLS side, which is used
in testing to force a rehandshake. We do not implement it on the OpenSSL
side, so stub it out: it throws an exception if called.

In the one test that uses it, we just skip actually forcing a
rehandshake when using OpenSSL (so the value of the test is reduced).

See also CORE-14961.