future: fix value class operator()
In 950b54ef we changed the value type of the function object when
invoked to a temporary so that operator() && would selected if it
exists. However, this wasn't working correctly for then()-alike
functions such as then_wrapped() or finally(), which was revealed by
testing.

This change adds move() in the required places at function invocation
to ensure that the function type is treated as an rvalue and the rvalue
overload is selected.

(cherry picked from commit 0316ddc9dc2255cfcb2772ea65292fd5284f1045)