http::request: remove content deprecation
http::request::content is deprecated upstream, with the idea that you
set the server into streaming mode and use the input_stream<> in
the request directly.

This is not a totally trivial change, so for now we want to just keep
using content as this functionality is the same as always, so we
remove the deprecation from our fork for now.

See also CORE-15051.